+++ CHANGELOG.md
@@ -21,6 +21,7 @@ filters and searching the nearest frame without any annotations (<https://github
 - MacOS users notes in CONTRIBUTING.md
 - Ability to prepare meta information manually (<https://github.com/openvinotoolkit/cvat/pull/2217>)
 - Ability to upload prepared meta information along with a video when creating a task (<https://github.com/openvinotoolkit/cvat/pull/2217>)
+- Optional chaining plugin for cvat-canvas and cvat-ui (<https://github.com/openvinotoolkit/cvat/pull/2249>)
 
 ### Changed
 - UI models (like DEXTR) were redesigned to be more interactive (<https://github.com/opencv/cvat/pull/2054>)